Add files via upload
This code is written in C language. 
It takes the names of the teams, players and ball-by-ball scores with the specified no. of overs by the user and displays the Team Scorecard at the end of the each inning and after the completion of both innings the Main Scorecard will be displayed.